Fixed probability curve function
git-svn-id: file:///home/vsekhar/vgp/vgp@325 cb3134f3-094e-0410-9174-ff1f5a1024ca

+++ trunk/vgp/util/random.hpp
@@ -19,7 +19,7 @@ struct SeededGenerator : GENERATOR {
 
 template <class GENERATOR>
 struct TimeSeededGenerator : GENERATOR {
-	TimeSeededGenerator() : GENERATOR(std::time(0)) {};
+	TimeSeededGenerator() : GENERATOR(static_cast<unsigned int>(std::time(0))) {};
 };
 
 typedef TimeSeededGenerator<boost::mt19937> time_seeded_mersenne_t;
@@ -36,7 +36,6 @@ struct RandomBool : RandomBool_base {
 		return RandomBool_base::operator()() < probability_of_true;
 	}
 };
-
 
 } // namespace util
 } // namespace vgp
